tools/gemini-slack-bot/requirements.txt
slack-bolt>=1.21.0
slack-sdk>=3.33.0
google-genai>=1.0.0
python-dotenv>=1.0.0
